Create and implement reusable JobOffers component
